Create function that return number of bits

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -1,7 +1,7 @@
 #include "main.h"
 
 /**
- * flip_bits - function that returns the number of bits you would need to flip to get from one number to another.
+ * flip_bits - function that returns the number.
  *
  * @n: unsigned long int param.
  * @m: unsigned long int param.
@@ -9,12 +9,12 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-unsigned value = n ^ m;
-unsigned i = 0;
+unsigned long int value = n ^ m;
+unsigned int i = 0;
 while (i > 0)
 {
 i++;
-i &= (i - 1); 
+i &= (i - 1);
 }
 return (i);
 }
